switch to factory pattern install and configure Gunicorn and fix circular import problem
