+++ app/api.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from controllers.client import client_controller
+from .controllers.client import client_controller
 
 api = Blueprint('api', __name__)
 api.register_blueprint(client_controller, url_prefix='/client')
