+++ app/requirements.txt
@@ -5,10 +5,12 @@ Flask==3.1.1
 flask-cors==6.0.0
 Flask-Migrate==4.1.0
 Flask-SQLAlchemy==3.1.1
+gunicorn==23.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.6
 Mako==1.3.10
 MarkupSafe==3.0.2
+packaging==25.0
 psycopg2==2.9.10
 python-dotenv==1.1.0
 SQLAlchemy==2.0.41
